+++ vcd/resource_vcd_lb_virtual_server_test.go
@@ -38,7 +38,7 @@ func TestAccVcdLbVirtualServer(t *testing.T) {
 		return
 	}
 
-	resource.Test(t, resource.TestCase{
+	resource.ParallelTest(t, resource.TestCase{
 		Providers:    testAccProviders,
 		PreCheck:     func() { testAccPreCheck(t) },
 		CheckDestroy: testAccCheckVcdLbVirtualServerDestroy(params["VirtualServerName"].(string)),