+++ vcd/resource_vcd_vapp_vm.go
@@ -115,20 +115,17 @@ func resourceVcdVAppVm() *schema.Resource {
 			},
 			"network": {
 				ConflictsWith: []string{"ip", "network_name", "vapp_network_name", "network_href"},
-				ForceNew:      true,
 				Optional:      true,
 				Type:          schema.TypeList,
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
 						"type": {
-							ForceNew:     true,
 							Required:     true,
 							Type:         schema.TypeString,
 							ValidateFunc: validation.StringInSlice([]string{"vapp", "org", "none"}, false),
 							Description:  "Network type to use: 'vapp', 'org' or 'none'. Use 'vapp' for vApp network, 'org' to attach Org VDC network. 'none' for empty NIC.",
 						},
 						"ip_allocation_mode": {
-							ForceNew:     true,
 							Optional:     true,
 							Type:         schema.TypeString,
 							ValidateFunc: validation.StringInSlice([]string{"POOL", "DHCP", "MANUAL", "NONE"}, false),
@@ -140,14 +137,12 @@ func resourceVcdVAppVm() *schema.Resource {
 						},
 						"ip": {
 							Computed:     true,
-							ForceNew:     true,
 							Optional:     true,
 							Type:         schema.TypeString,
 							ValidateFunc: checkEmptyOrSingleIP(), // Must accept empty string to ease using HCL interpolation
 						},
 						"is_primary": {
 							Default:  false,
-							ForceNew: true,
 							Optional: true,
 							// By default if the value is omitted it will report schema change
 							// on every terraform operation. The below function
@@ -202,7 +197,46 @@ func resourceVcdVAppVm() *schema.Resource {
 				Default:     false,
 				Description: "Expose hardware-assisted CPU virtualization to guest OS.",
 			},
+			"guest_properties": {
+				Type:        schema.TypeMap,
+				Optional:    true,
+				Description: "Key/value settings for guest properties",
+			},
+			"customization": &schema.Schema{
+				Optional:    true,
+				MinItems:    1,
+				MaxItems:    1,
+				Type:        schema.TypeList,
+				Description: "Guest customization block",
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"force": {
+							ValidateFunc: noopValueWarningValidator(true,
+								"Using 'true' value for field 'vcd_vapp_vm.customization.force' will reboot VM on every 'terraform apply' operation"),
+							Type:     schema.TypeBool,
+							Optional: true,
+							Default:  false,
+							// This settings is used as a 'flag' and it does not matter what is set in the
+							// state. If it is 'true' - then it means that 'update' procedure must set the
+							// VM for customization at next boot and reboot it.
+							DiffSuppressFunc: suppressFalse(),
+						},
+					},
+				},
+			},
 		},
+	}
+}
+
+// noopValueWarningValidator is a no-op validator which only emits warning string when fieldValue
+// is set to the specified one
+func noopValueWarningValidator(fieldValue interface{}, warningText string) schema.SchemaValidateFunc {
+	return func(i interface{}, k string) (warnings []string, errors []error) {
+		if fieldValue == i {
+			warnings = append(warnings, fmt.Sprintf("%s\n\n", warningText))
+		}
+
+		return
 	}
 }
 
@@ -246,6 +280,13 @@ func falseBoolSuppress() schema.SchemaDiffSuppressFunc {
 	}
 }
 
+// suppressNewFalse always suppresses when new value is false
+func suppressFalse() schema.SchemaDiffSuppressFunc {
+	return func(k string, old string, new string, d *schema.ResourceData) bool {
+		return new == "false"
+	}
+}
+
 func resourceVcdVAppVmCreate(d *schema.ResourceData, meta interface{}) error {
 	vcdClient := meta.(*VCDClient)
 
@@ -283,10 +324,10 @@ func resourceVcdVAppVmCreate(d *schema.ResourceData, meta interface{}) error {
 	// TODO v3.0 remove else branch once 'network_name', 'vapp_network_name', 'ip' are deprecated
 	networkConnectionSection := types.NetworkConnectionSection{}
 	if len(d.Get("network").([]interface{})) > 0 {
-		networkConnectionSection, err = networksToConfig(d.Get("network").([]interface{}), *vdc, vapp, vcdClient)
+		networkConnectionSection, err = networksToConfig(d.Get("network").([]interface{}), vdc, vapp, vcdClient)
 	} else {
 		networkConnectionSection, err = deprecatedNetworksToConfig(d.Get("network_name").(string),
-			d.Get("vapp_network_name").(string), d.Get("ip").(string), *vdc, vapp, vcdClient)
+			d.Get("vapp_network_name").(string), d.Get("ip").(string), vdc, vapp, vcdClient)
 	}
 	if err != nil {
 		return fmt.Errorf("unable to process network configuration: %s", err)
@@ -335,12 +376,25 @@ func resourceVcdVAppVmCreate(d *schema.ResourceData, meta interface{}) error {
 		}
 	}
 
+	if _, ok := d.GetOk("guest_properties"); ok {
+		vmProperties, err := getGuestProperties(d)
+		if err != nil {
+			return fmt.Errorf("unable to convert guest properties to data structure")
+		}
+
+		log.Printf("[TRACE] Setting VM guest properties")
+		_, err = vm.SetProductSectionList(vmProperties)
+		if err != nil {
+			return fmt.Errorf("error setting guest properties: %s", err)
+		}
+	}
+
 	d.SetId(d.Get("name").(string))
 
 	// TODO do not trigger resourceVcdVAppVmUpdate from create. These must be separate actions.
 	err = resourceVcdVAppVmUpdateExecute(d, meta)
 	if err != nil {
-		errAttachedDisk := updateStateOfAttachedDisks(d, vm, *vdc)
+		errAttachedDisk := updateStateOfAttachedDisks(d, vm, vdc)
 		if errAttachedDisk != nil {
 			d.Set("disk", nil)
 			return fmt.Errorf("error reading attached disks : %#v and internal error : %#v", errAttachedDisk, err)
@@ -352,7 +406,7 @@ func resourceVcdVAppVmCreate(d *schema.ResourceData, meta interface{}) error {
 
 // Adds existing org VDC network to VM network configuration
 // Returns configured OrgVDCNetwork for Vm, networkName, error if any occur
-func addVdcNetwork(networkNameToAdd string, vdc govcd.Vdc, vapp govcd.VApp, vcdClient *VCDClient) (*types.OrgVDCNetwork, error) {
+func addVdcNetwork(networkNameToAdd string, vdc *govcd.Vdc, vapp govcd.VApp, vcdClient *VCDClient) (*types.OrgVDCNetwork, error) {
 	if networkNameToAdd == "" {
 		return &types.OrgVDCNetwork{}, fmt.Errorf("'network_name' must be valid when adding VM to raw vApp")
 	}
@@ -500,10 +554,25 @@ func resourceVcdVAppVmUpdateExecute(d *schema.ResourceData, meta interface{}) er
 		return fmt.Errorf("error getting VM2: %#v", err)
 	}
 
-	status, err := vm.GetStatus()
+	vmStatusBeforeUpdate, err := vm.GetStatus()
 	if err != nil {
-		return fmt.Errorf("error getting VM status: %#v", err)
+		return fmt.Errorf("error getting VM status before update: %#v", err)
 	}
+
+	if d.HasChange("guest_properties") {
+		vmProperties, err := getGuestProperties(d)
+		if err != nil {
+			return fmt.Errorf("unable to convert guest properties to data structure")
+		}
+
+		log.Printf("[TRACE] Updating VM guest properties")
+		_, err = vm.SetProductSectionList(vmProperties)
+		if err != nil {
+			return fmt.Errorf("error setting guest properties: %s", err)
+		}
+	}
+	// Check if the user requested for forced customization of VM
+	customizationNeeded := isForcedCustomization(d.Get("customization"))
 
 	// VM does not have to be in POWERED_OFF state for metadata operations
 	if d.HasChange("metadata") {
@@ -541,8 +610,16 @@ func resourceVcdVAppVmUpdateExecute(d *schema.ResourceData, meta interface{}) er
 	}
 
 	if d.HasChange("memory") || d.HasChange("cpus") || d.HasChange("cpu_cores") || d.HasChange("power_on") || d.HasChange("disk") ||
-		d.HasChange("expose_hardware_virtualization") {
-		if status != "POWERED_OFF" {
+		d.HasChange("expose_hardware_virtualization") || d.HasChange("network") {
+
+		log.Printf("[TRACE] VM %s has changes: memory(%t), cpus(%t), cpu_cores(%t), power_on(%t), disk(%t), expose_hardware_virtualization(%t), network(%t)",
+			vm.VM.Name, d.HasChange("memory"), d.HasChange("cpus"), d.HasChange("cpu_cores"), d.HasChange("power_on"), d.HasChange("disk"),
+			d.HasChange("expose_hardware_virtualization"), d.HasChange("network"))
+
+		// If customization is not requested then a simple shutdown is enough
+		if vmStatusBeforeUpdate != "POWERED_OFF" && !customizationNeeded {
+			log.Printf("[DEBUG] Powering off VM %s for offline update. Previous state %s",
+				vm.VM.Name, vmStatusBeforeUpdate)
 			task, err := vm.PowerOff()
 			if err != nil {
 				return fmt.Errorf("error Powering Off: %#v", err)
@@ -553,11 +630,25 @@ func resourceVcdVAppVmUpdateExecute(d *schema.ResourceData, meta interface{}) er
 			}
 		}
 
+		// If customization was requested then a shutdown with undeploy is needed
+		if vmStatusBeforeUpdate != "POWERED_OFF" && customizationNeeded {
+			log.Printf("[DEBUG] Un-deploying VM %s for offline update. Previous state %s",
+				vm.VM.Name, vmStatusBeforeUpdate)
+			task, err := vm.Undeploy()
+			if err != nil {
+				return fmt.Errorf("error triggering undeploy for VM %s: %s", vm.VM.Name, err)
+			}
+			err = task.WaitTaskCompletion()
+			if err != nil {
+				return fmt.Errorf("error waiting for undeploy task for VM %s: %s", vm.VM.Name, err)
+			}
+		}
+
 		// detaching independent disks - only possible when VM power off
 		if d.HasChange("disk") {
-			err = attachDetachDisks(d, vm, *vdc)
+			err = attachDetachDisks(d, vm, vdc)
 			if err != nil {
-				errAttachedDisk := updateStateOfAttachedDisks(d, vm, *vdc)
+				errAttachedDisk := updateStateOfAttachedDisks(d, vm, vdc)
 				if errAttachedDisk != nil {
 					d.Set("disk", nil)
 					return fmt.Errorf("error reading attached disks : %#v and internal error : %#v", errAttachedDisk, err)
@@ -611,26 +702,69 @@ func resourceVcdVAppVmUpdateExecute(d *schema.ResourceData, meta interface{}) er
 			}
 		}
 
-		if d.Get("power_on").(bool) {
+		if d.HasChange("network") {
+			networkConnectionSection, err := networksToConfig(d.Get("network").([]interface{}), vdc, vapp, vcdClient)
+			if err != nil {
+				return fmt.Errorf("unable to setup network configuration for update: %s", err)
+			}
+			err = vm.UpdateNetworkConnectionSection(&networkConnectionSection)
+			if err != nil {
+				return fmt.Errorf("unable to update network configuration: %s", err)
+			}
+		}
 
+	}
+
+	// If the VM was powered off during update but it has to be powered off
+	if d.Get("power_on").(bool) {
+		vmStatus, err := vm.GetStatus()
+		if err != nil {
+			return fmt.Errorf("error getting VM status before ensuring it is powered on: %s", err)
+		}
+		log.Printf("[DEBUG] Powering on VM %s after update. Previous state %s", vm.VM.Name, vmStatus)
+
+		// Simply power on if customization is not requested
+		if !customizationNeeded && vmStatus != "POWERED_ON" {
 			task, err := vm.PowerOn()
 			if err != nil {
-				return fmt.Errorf("error Powering Up: %#v", err)
+				return fmt.Errorf("error powering on: %s", err)
 			}
-
 			err = task.WaitTaskCompletion()
 			if err != nil {
 				return fmt.Errorf(errorCompletingTask, err)
 			}
 		}
 
+		// When customization is requested VM must be un-deployed before starting it
+		if customizationNeeded {
+			log.Printf("[TRACE] forced customization for VM %s was requested. Current state %s",
+				vm.VM.Name, vmStatus)
+
+			if vmStatus != "POWERED_OFF" {
+				log.Printf("[TRACE] VM %s is in state %s. Un-deploying", vm.VM.Name, vmStatus)
+				task, err := vm.Undeploy()
+				if err != nil {
+					return fmt.Errorf("error triggering undeploy for VM %s: %s", vm.VM.Name, err)
+				}
+				err = task.WaitTaskCompletion()
+				if err != nil {
+					return fmt.Errorf("error waiting for undeploy task for VM %s: %s", vm.VM.Name, err)
+				}
+			}
+
+			log.Printf("[TRACE] Powering on VM %s with forced customization", vm.VM.Name)
+			err = vm.PowerOnAndForceCustomization()
+			if err != nil {
+				return fmt.Errorf("failed powering on with customization: %s", err)
+			}
+		}
 	}
 
 	return resourceVcdVAppVmRead(d, meta)
 }
 
 // updates attached disks to latest state. Removed not needed and add new ones
-func attachDetachDisks(d *schema.ResourceData, vm govcd.VM, vdc govcd.Vdc) error {
+func attachDetachDisks(d *schema.ResourceData, vm govcd.VM, vdc *govcd.Vdc) error {
 	oldValues, newValues := d.GetChange("disk")
 
 	attachDisks := newValues.(*schema.Set).Difference(oldValues.(*schema.Set))
@@ -749,7 +883,18 @@ func resourceVcdVAppVmRead(d *schema.ResourceData, meta interface{}) error {
 	d.Set("href", vm.VM.HREF)
 	d.Set("expose_hardware_virtualization", vm.VM.NestedHypervisorEnabled)
 
-	err = updateStateOfAttachedDisks(d, vm, *vdc)
+	// update guest properties
+	guestProperties, err := vm.GetProductSectionList()
+	if err != nil {
+		return fmt.Errorf("unable to read guest properties: %s", err)
+	}
+
+	err = setGuestProperties(d, guestProperties)
+	if err != nil {
+		return fmt.Errorf("unable to set guest properties in state: %s", err)
+	}
+
+	err = updateStateOfAttachedDisks(d, vm, vdc)
 	if err != nil {
 		d.Set("disk", nil)
 		return fmt.Errorf("error reading attached disks : %#v", err)
@@ -758,7 +903,7 @@ func resourceVcdVAppVmRead(d *schema.ResourceData, meta interface{}) error {
 	return nil
 }
 
-func updateStateOfAttachedDisks(d *schema.ResourceData, vm govcd.VM, vdc govcd.Vdc) error {
+func updateStateOfAttachedDisks(d *schema.ResourceData, vm govcd.VM, vdc *govcd.Vdc) error {
 	// Check VM independent disks state
 	diskProperties, err := expandDisksProperties(d.Get("disk"))
 	if err != nil {
@@ -830,7 +975,7 @@ func resourceVcdVAppVmDelete(d *schema.ResourceData, meta interface{}) error {
 		return fmt.Errorf("error getting VM status: %#v", err)
 	}
 
-	log.Printf("[TRACE] VM Status:: %s", status)
+	log.Printf("[TRACE] VM Status: %s", status)
 	if status != "POWERED_OFF" {
 		log.Printf("[TRACE] Undeploying VM: %s", vm.VM.Name)
 		task, err := vm.Undeploy()
@@ -865,6 +1010,7 @@ func resourceVcdVAppVmDelete(d *schema.ResourceData, meta interface{}) error {
 	}
 
 	log.Printf("[TRACE] Removing VM: %s", vm.VM.Name)
+
 	err = vapp.RemoveVM(vm)
 	if err != nil {
 		return fmt.Errorf("error deleting: %#v", err)
@@ -891,7 +1037,7 @@ func resourceVcdVmIndependentDiskHash(v interface{}) int {
 
 // networksToConfig converts terraform schema for 'networks' and converts to types.NetworkConnectionSection
 // which is used for creating new VM
-func networksToConfig(networks []interface{}, vdc govcd.Vdc, vapp govcd.VApp, vcdClient *VCDClient) (types.NetworkConnectionSection, error) {
+func networksToConfig(networks []interface{}, vdc *govcd.Vdc, vapp govcd.VApp, vcdClient *VCDClient) (types.NetworkConnectionSection, error) {
 	networkConnectionSection := types.NetworkConnectionSection{}
 	for index, singleNetwork := range networks {
 		nic := singleNetwork.(map[string]interface{})
@@ -942,7 +1088,7 @@ func networksToConfig(networks []interface{}, vdc govcd.Vdc, vapp govcd.VApp, vc
 
 // deprecatedNetworksToConfig converts deprecated network configuration in fields
 // TODO v3.0 remove this function once 'network_name', 'vapp_network_name', 'ip' are deprecated
-func deprecatedNetworksToConfig(network_name, vapp_network_name, ip string, vdc govcd.Vdc, vapp govcd.VApp, vcdClient *VCDClient) (types.NetworkConnectionSection, error) {
+func deprecatedNetworksToConfig(network_name, vapp_network_name, ip string, vdc *govcd.Vdc, vapp govcd.VApp, vcdClient *VCDClient) (types.NetworkConnectionSection, error) {
 	if vapp_network_name != "" {
 		isVappNetwork, err := isItVappNetwork(vapp_network_name, vapp)
 		if err != nil {
@@ -1085,4 +1231,71 @@ func readNetworks(vm govcd.VM, vapp govcd.VApp) ([]map[string]interface{}, error
 		nets = append(nets, singleNIC)
 	}
 	return nets, nil
+}
+
+// isForcedCustomization checks "customization" block in resource and checks if the value of field "force"
+// is set to "true". It returns false if the value is not set or is set to false
+func isForcedCustomization(customizationBlock interface{}) bool {
+	customizationSlice := customizationBlock.([]interface{})
+
+	if len(customizationSlice) != 1 {
+		return false
+	}
+
+	cust := customizationSlice[0]
+	fc := cust.(map[string]interface{})
+	forceCust, ok := fc["force"]
+	forceCustBool := forceCust.(bool)
+
+	if !ok || !forceCustBool {
+		return false
+	}
+
+	return true
+}
+
+// getGuestProperties returns a struct for setting guest properties
+func getGuestProperties(d *schema.ResourceData) (*types.ProductSectionList, error) {
+	guestProperties := d.Get("guest_properties")
+	guestProp := convertToStringMap(guestProperties.(map[string]interface{}))
+	vmProperties := &types.ProductSectionList{
+		ProductSection: &types.ProductSection{
+			Info:     "Custom properties",
+			Property: []*types.Property{},
+		},
+	}
+	for key, value := range guestProp {
+		log.Printf("[TRACE] Adding guest property: key=%s, value=%s to object", key, value)
+		oneProp := &types.Property{
+			UserConfigurable: true,
+			Type:             "string",
+			Key:              key,
+			Label:            key,
+			Value:            &types.Value{Value: value},
+		}
+		vmProperties.ProductSection.Property = append(vmProperties.ProductSection.Property, oneProp)
+	}
+
+	return vmProperties, nil
+}
+
+// setGuestProperties sets guest properties into state
+func setGuestProperties(d *schema.ResourceData, properties *types.ProductSectionList) error {
+	data := make(map[string]string)
+
+	// if properties object does not have actual properties - set state to empty
+	log.Printf("[TRACE] Setting empty properties into statefile because no properties were specified")
+	if properties == nil || properties.ProductSection == nil || len(properties.ProductSection.Property) == 0 {
+		return d.Set("guest_properties", make(map[string]string))
+	}
+
+	for _, prop := range properties.ProductSection.Property {
+		// if a value was set - use it
+		if prop.Value != nil {
+			data[prop.Key] = prop.Value.Value
+		}
+	}
+
+	log.Printf("[TRACE] Setting properties into statefile")
+	return d.Set("guest_properties", data)
 }