+++ CHANGELOG.md
@@ -18,12 +18,17 @@ IMPROVEMENTS:
 * `resource/vcd_catalog_item`: CatalogItem state ID changed from colon separated list of catalog name and item name to vCD ID 
 * `resource/catalog_item` added catalog item metadata support [#298] 
 * `resource/catalog_media` added catalog media item metadata support [#298]
+* `resource/vcd_vapp_vm` supports update for `network` block [#310]
+* `resource/vcd_vapp_vm` allows to force guest customization [#310]
+* `resource/vcd_vapp` supports guest properties [#319]
+* `resource/vcd_vapp_vm` supports guest properties [#319]
 * Upgrade Terraform SDK dependency to 0.12.6 [#302]
 
 BUG FIXES:
 * Change default value for `vcd_org.deployed_vm_quota` and `vcd_org.stored_vm_quota`. It was incorrectly set at `-1` instead of `0`.
 * Change Org ID from partial task ID to real Org ID during creation.
 * Wait for task completion on creation and update, where tasks were not handled at all.
+* `resource/vcd_firewall_rules` force recreation of the resource when attributes of the sub-element `rule` are changed (fixes a situation when it tried to update a rule).
 
 ## 2.4.0 (July 29, 2019)